Module Tezt_etherlink.Block

type transactions =

| Hash of string list

| Full of Transaction.transaction_object list

| Empty

Blocks can contain either only hashes or full transaction objects. As the empty list case cannot be discriminated between the two types, `Empty` acts as a wildcard for the empty transaction list.

type t = {

number : int32;

hash : string;

parent : string;

nonce : string;

sha3Uncles : string;

logsBloom : string option;

transactionRoot : string;

stateRoot : string;

receiptRoot : string;

miner : string;

difficulty : int64;

totalDifficulty : int64;

extraData : string;

size : int32;

gasLimit : int64;

gasUsed : int64;

timestamp : Tezos_base.Time.Protocol.t;

transactions : transactions;

uncles : string list;

baseFeePerGas : int64;

prevRandao : string;

}

Simplified Ethereum block representation.

val of_json : Tezt_wrapper.JSON.t -> t

Extracts a block t from a JSON.