Module Tezt_etherlink.Durable_storage_path

val no_0x : string -> string

no_0x s removes the prefix 0x of s if it exists.

val normalize : string -> string

normalize s calls no_0x and String.lowercase_ascii on s.

type path = string

Path in the Wasm PVM durable storage.

val delayed_inbox : path

val kernel_root_hash : path

kernel_root_hash is the path to the current kernel root hash.

val indexes : path

indexes is the directory with all indexes category.

val eth_accounts : path

eth_accounts is the path to ethereum accounts.

val eth_account : string -> path

eth_account addr is the path to the addr account. The address is "normalized", i.e. lowered and removed the prefix 0x if it exists.

val balance : string -> path

balance addr is the path to the addr account's balance.

val code : string -> path

code addr is the path to the addr account's code.

val storage : string -> ?key:string -> unit -> path

storage addr ?key () is the path to the addr storage's code. key can be provided to get the path of a sub-element in the storage.

val admin : path

admin is the path to the administrator contract.

val kernel_governance : path

kernel_governance is the path to the kernel governance contract.

val kernel_security_governance : path

kernel_security_governance is the path to the security governance contract.

val sequencer_governance : path

sequencer_governance is the path to the governance contract administrating the sequencer.

val ticketer : path

ticketer is the path to the ticketer contract.

val sequencer : path

sequencer is the path to the sequencer flag.

val sequencer_pool_address : path

sequencer_pool_address is the path to the L2 address credited with DA fees.

val kernel_boot_wasm : path

kernel_boot_wasm is the path to the kernel `boot.wasm`.

val delayed_bridge_path : path

delayed_bridge_path is the path to the delayed transaction bridge contract.

val da_fee_per_byte_path : path

da_fee_per_byte_path is the path to the da fee per byte, charged on every transaction.

val minimum_base_fee_per_gas : path

minimum_base_fee_per_gas is the path to the minimum base fee per gas the kernel will charge.

val delayed_inbox_timeout : path

delayed_inbox_timeout is the path to the timeout for delayed transactions.

val delayed_inbox_min_levels : path

delayed_inbox_min_levels is the path to the minimum number of L1 levels needed to have passed to consider a timeout.

val reveal_config : path

reveal_config is the path to the storage configuration.

val enable_fa_bridge : path

enable_fa_bridge is the path to the feature flag to activate the FA bridge.

module Ticket_table : sig ... end

module Ghostnet : sig ... end