Module Tezt_etherlink.Evm_node

type t

EVM node server state.

type time_between_blocks =

| Nothing
(*
Does not produce any block if not forced by the private RPC
*)

| Time_between_blocks of float
(*
Interval at which the sequencer creates an empty block by default.
*)

type mode =

| Observer of {

initial_kernel : string;

preimages_dir : string option;

private_rpc_port : int option;
(*
Port for private RPC server
*)

rollup_node_endpoint : string;

}

| Threshold_encryption_observer of {

initial_kernel : string;

preimages_dir : string;

rollup_node_endpoint : string;

bundler_node_endpoint : string;

}

| Sequencer of {

initial_kernel : string;
(*
Path to the initial kernel used by the sequencer.
*)

preimage_dir : string option;
(*
Path to the directory with the associated preimages.
*)

private_rpc_port : int option;
(*
Port for private RPC server
*)

time_between_blocks : time_between_blocks option;
(*
See time_between_blocks, if the value is not provided, the sequencer uses it default value.
*)

sequencer : string;
(*
Secret key used to sign the blueprints.
*)

genesis_timestamp : Tezt_tezos.Client.timestamp option;
(*
Genesis timestamp
*)

max_blueprints_lag : int option;

max_blueprints_ahead : int option;

max_blueprints_catchup : int option;

catchup_cooldown : int option;

max_number_of_chunks : int option;

wallet_dir : string option;
(*
--wallet-dir: client directory.
*)

tx_pool_timeout_limit : int option;
(*
--tx-pool-timeout-limit: transaction timeout inside the pool.
*)

tx_pool_addr_limit : int option;
(*
--tx-pool-addr-limit: maximum address allowed simultaneously inside the pool.
*)

tx_pool_tx_per_addr_limit : int option;
(*
--tx-pool-tx-per-addr-limit: maximum transaction per address allowed simultaneously inside the pool.
*)

dal_slots : int list option;

}

| Sandbox of {

initial_kernel : string;

preimage_dir : string option;

private_rpc_port : int option;

time_between_blocks : time_between_blocks option;

genesis_timestamp : Tezt_tezos.Client.timestamp option;

max_number_of_chunks : int option;

wallet_dir : string option;

tx_pool_timeout_limit : int option;

tx_pool_addr_limit : int option;

tx_pool_tx_per_addr_limit : int option;

}

| Threshold_encryption_sequencer of {

initial_kernel : string;
(*
Path to the initial kernel used by the sequencer.
*)

preimage_dir : string option;
(*
Path to the directory with the associated preimages.
*)

private_rpc_port : int option;
(*
Port for private RPC server
*)

time_between_blocks : time_between_blocks option;
(*
See time_between_blocks, if the value is not provided, the sequencer uses it default value.
*)

sequencer : string;
(*
Secret key used to sign the blueprints.
*)

genesis_timestamp : Tezt_tezos.Client.timestamp option;
(*
Genesis timestamp
*)

max_blueprints_lag : int option;

max_blueprints_ahead : int option;

max_blueprints_catchup : int option;

catchup_cooldown : int option;

max_number_of_chunks : int option;

wallet_dir : string option;
(*
--wallet-dir: client directory.
*)

tx_pool_timeout_limit : int option;
(*
--tx-pool-timeout-limit: transaction timeout inside the pool.
*)

tx_pool_addr_limit : int option;
(*
--tx-pool-addr-limit: maximum address allowed simultaneously inside the pool.
*)

tx_pool_tx_per_addr_limit : int option;
(*
--tx-pool-tx-per-addr-limit: maximum transaction per address allowed simultaneously inside the pool.
*)

sequencer_sidecar_endpoint : string;
(*
--sequencer-sidecar-endpoint: Uri of the sidecar endpoints to which proposals are forwarded, and from where preblocks are fetched.
*)

dal_slots : int list option;

}

| Proxy

| Rpc of mode

EVM node mode.

val mode : t -> mode

Returns the mode of the EVM node.

val name : t -> string

Returns the name of the EVM node.

val data_dir : t -> string

Returns the data_dir of the EVM node.

val preimages_dir : t -> string

Returns the path to the directory storing the preimages used by the kernel runned by the node.

val supports_threshold_encryption : t -> bool

val create : ?path:string -> ?name:string -> ?runner:Tezt_wrapper.Runner.t -> ?mode:mode -> ?data_dir:string -> ?rpc_addr:string -> ?rpc_port:int -> ?restricted_rpcs:string -> ?websockets:bool -> string -> t

create ?name ?runner ?mode ?data_dir ?rpc_addr ?rpc_port rollup_node_endpoint creates an EVM node server.

The server listens to requests at address rpc_addr and the port rpc_port. rpc_addr defaults to Constant.default_host and a fresh port is chosen if rpc_port is not set. The EVM node starts a websocket server for JSON-RPC communication if websocket is true.

The server communicates with a rollup-node and sets its endpoint via rollup_node_endpoint.

mode defaults to Proxy.

val initial_kernel : t -> string

initial_kernel node returns the path to the kernel used to initialize the EVM state. Fails if node is a proxy node.

val run : ?wait:bool -> ?extra_arguments:string list -> t -> unit Lwt.t

run ?wait ?extra_arguments evm_node launches the EVM node server with the arguments given during create, additional arguments can be passed via extra_arguments. wait defaults to true, if it is set to false, the evm node is ran but we do not wait for it to be ready.

val wait_for_event : ?timeout:float -> t -> event:string -> (Tezt_wrapper.JSON.t -> 'a option) -> 'a Lwt.t

wait_for_event ?timeout evm_node ~event f waits for event event until timeout on node evm_node.

The wait is resolved if an event with name event is seen and f returns Some _.

val wait_for_ready : ?timeout:float -> t -> unit Lwt.t

wait_for_ready evm_node waits until evm_node is ready.

val wait_for_blueprint_applied : ?timeout:float -> t -> int -> unit Lwt.t

wait_for_blueprint_applied ~timeout evm_node level waits until evm_node has applied a blueprint locally for level level.

val wait_for_blueprint_invalid_applied : t -> unit Lwt.t

wait_for_blueprint_invalid_applied waits for the event blueprint_invalid_applied.v0.

val wait_for_blueprint_finalized : ?timeout:float -> t -> int -> unit Lwt.t

val wait_for_predownload_kernel : ?timeout:float -> t -> root_hash:string -> unit Lwt.t

wait_for_predownload_kernel ?timeout evm_node ~root_hash waits until evm_node has download a kernel with root_hash.

val wait_for_blueprint_invalid : ?timeout:float -> t -> unit Lwt.t

wait_for_blueprint_invalid ~timeout evm_node waits until evm_node has seen an invalid blueprint.

val wait_for_blueprint_injected : ?timeout:float -> t -> int -> unit Lwt.t

wait_for_blueprint_injected ~timeout evm_node level waits until evm_node has injected a blueprint for level level to its rollup node.

val wait_for_blueprint_injected_on_dal : ?timeout:float -> t -> (int * int) Lwt.t

wait_for_blueprint_injected_on_dal ~timeout evm_node level waits until evm_node has injected a blueprint to the DAL node and returns the L1 level of injection and the number of chunks injected. Can be used with wait_for_blueprint_injected to get the expected level.

val wait_for_signal_signed : ?timeout:float -> t -> (string * (int * int) list) Lwt.t

wait_for_signal_signed ~timeout evm_node waits until (at least) a signal has been signed and returns the smart rollup address for which the signal has been signed together with the slot indices and the published levels that have been signaled.

val wait_for_pending_upgrade : ?timeout:float -> t -> (string * string) Lwt.t

val wait_for_successful_upgrade : ?timeout:float -> t -> (string * int) Lwt.t

val wait_for_block_producer_locked : ?timeout:float -> t -> unit Lwt.t

val wait_for_block_producer_tx_injected : ?timeout:float -> t -> string Lwt.t

val wait_for_retrying_connect : ?timeout:float -> t -> unit Lwt.t

val wait_for_flush_delayed_inbox : ?timeout:float -> ?level:int -> t -> int Lwt.t

val wait_for_rollup_node_follower_connection_acquired : ?timeout:float -> t -> unit Lwt.t

val wait_for_rollup_node_follower_disabled : ?timeout:float -> t -> unit Lwt.t

type processed_l1_level = {

l1_level : int;

finalized_blueprint : int;

}

val wait_for_processed_l1_level : ?timeout:float -> ?level:int -> t -> processed_l1_level Lwt.t

wait_for_processed_l1_level ?timeout ?level evm_node waits until a (finalized) layer1 level has been processed and returns the layer1 level and it's latest corresponding blueprints. If level = Some level then it waits until that specific layer1 level is processed. It returns the tuple (layer1_level, finalized_blueprint).

val wait_for_blueprint_catchup : ?timeout:float -> t -> (int * int) Lwt.t

val wait_for_blueprint_injection_failure : ?timeout:float -> ?level:int -> t -> unit Lwt.t

module Config_file : sig ... end

Node configuration files.

val spawn_init_config : ?extra_arguments:string list -> t -> Tezt_wrapper.Process.t

spawn_init_config ?extra_arguments evm_node runs "init config" with arguments found in the state.

val spawn_init_config_minimal : data_dir:string -> ?path:string -> ?extra_arguments:string list -> unit -> Tezt_wrapper.Process.t

spawn_init_config_minimal ~data_dir ?path ?extra_arguments () creates a minimal config with no cli argument populated as spawn_init_config.

Unlike spawn_init_config, does not require a Evm_node.t instance.

type garbage_collector_parameters = {

split_frequency_in_seconds : int;

number_of_chunks : int;

}

type history_mode =

| Archive

| Rolling

type rpc_server =

| Resto

| Dream

val patch_config_with_experimental_feature : ?drop_duplicate_when_injection:bool -> ?blueprints_publisher_order_enabled:bool -> ?block_storage_sqlite3:bool -> ?next_wasm_runtime:bool -> ?garbage_collector_parameters:garbage_collector_parameters -> ?history_mode:history_mode -> ?rpc_server:rpc_server -> ?enable_websocket:bool -> ?max_websocket_message_length:int -> unit -> Tezt_wrapper.JSON.t -> Tezt_wrapper.JSON.t

patch_config_with_experimental_feature ?drop_duplicate_when_injection ?block_storage_sqlite3 ?next_wasm_runtime ?rpc_server ?enable_websocket ?max_websocket_message_length json_config patches a config to add experimental feature. Each optional argument add the correspondent experimental feature.

val init : ?patch_config:(Tezt_wrapper.JSON.t -> Tezt_wrapper.JSON.t) -> ?name:string -> ?runner:Tezt_wrapper.Runner.t -> ?mode:mode -> ?data_dir:string -> ?rpc_addr:string -> ?rpc_port:int -> ?restricted_rpcs:string -> ?websockets:bool -> string -> t Lwt.t

init ?patch_config ?name ?runner ?mode ?data_dir ?rpc_addr ?rpc_port rollup_node_endpoint creates an EVM node server with create, init the config with spawn_init_config, patch it with patch_config, then runs it with run.

val rpc_port : t -> int

Get the RPC port given as --rpc-port to a node.

val spawn_run : ?extra_arguments:string list -> t -> Tezt_wrapper.Process.t

spawn_run ?extra_arguments evm_node same as run but spawns a process.

val terminate : ?timeout:float -> t -> unit Lwt.t

Send SIGTERM and wait for the process to terminate.

Default timeout is 30 seconds, after which SIGKILL is sent.

val resolve_or_timeout : ?timeout:float -> t -> name:string -> 'a Lwt.t -> 'a Lwt.t

val wait_for : ?where:string -> t -> string -> (Tezt_wrapper.JSON.t -> 'a option) -> 'a Lwt.t

The same exact behavior as Sc_rollup_node.wait_for but for the EVM node.

type delayed_transaction_kind =

| Deposit

| Transaction

| FaDeposit

type 'a evm_event_kind =

| Kernel_upgrade : (string * Tezt_tezos.Client.Time.t) evm_event_kind

| Sequencer_upgrade : (string * Hex.t * Tezt_tezos.Client.Time.t) evm_event_kind

| Blueprint_applied : (int * string) evm_event_kind

| New_delayed_transaction : (delayed_transaction_kind * string) evm_event_kind

val wait_for_evm_event : ?timeout:float -> 'a evm_event_kind -> ?check:(Tezt_wrapper.JSON.t -> 'a option) -> t -> 'a Lwt.t

wait_for_evm_event evm_node ~event_kind wait for the event evm_events_new_event.v0 using wait_for where the event kind is equal to event_kind (e.g. "sequencer_upgrade").

val wait_for_diverged : t -> (int * string * string) Lwt.t

wait_for_diverged evm_node waits for the event evm_events_follower_diverged.v0 using wait_for and return the diverging blueprint level expected hash, and found hash.

val wait_for_reset : t -> unit Lwt.t

wait_for_reset evm_node waits for the event evm_context_reset_at_level using wait_for. It does not wait for a specific level.

val wait_for_missing_blueprint : t -> (int * string) Lwt.t

wait_for_missing_blueprint evm_node waits for the event evm_events_follower_missing_blueprint.v0 using wait_for and return the missing blueprint level and block hash.

val wait_for_rollup_node_ahead : t -> int Lwt.t

wait_for_rollup_node_ahead evm_node waits for the event evm_events_follower_rollup_node_ahead.v0 using wait_for and returns the missing blueprint level.

val wait_for_tx_pool_add_transaction : ?timeout:float -> t -> string Lwt.t

wait_for_tx_pool_add_transaction ?timeout evm_node waits for the event tx_pool_add_transaction.v0 using wait_for and returns the transaction hash.

val wait_for_shutdown_event : ?can_terminate:bool -> t -> int option Lwt.t

wait_for_shutdown ?can_terminate evm_node waits until a node terminates and return its status. If the node is not running, make the test fail. If can_terminate is `true` and the node was already terminated, returns `None`.

val wait_for_split : ?level:int -> t -> int Lwt.t

wait_for_split ?level evm_node waits untils the node terminates splitting its irmin context at level level if provided.

val wait_for_gc_finished : ?gc_level:int -> ?head_level:int -> t -> (int * int) Lwt.t

wait_for_gc_finished ?gc_level ?head_level evm_node waits until the node terminates garbage collecting its context on head head_level at gc level gc_level if provided.

val wait_for_start_history_mode : ?history_mode:string -> t -> string Lwt.t

wait_for_start_history_mode ?history_mode evm_node waits until the start history mode event is emitted with history_mode if provided.

val rpc_endpoint : ?local:bool -> ?private_:bool -> t -> string

rpc_endpoint ?local ?private_ evm_node returns the endpoint to communicate with the evm_node. If private_ is true, the endpoint for the private RPC server is returned.

If local is given (false by default), then Constant.default_host is used (it overrides rpc-addr or the runner argument).

val endpoint : ?private_:bool -> t -> string

A deprecated alias for rpc_endpoint where local optional parameter is not given.

type request = {

method_ : string;

parameters : Tezt_wrapper.JSON.u;

}

JSON-RPC request.

val call_evm_rpc : ?private_:bool -> t -> request -> Tezt_wrapper.JSON.t Lwt.t

call_evm_rpc ?private_ evm_node request sends a JSON-RPC request to the evm_node, for the given request. If private_ is true, the request is sent to the private RPC server.

val batch_evm_rpc : ?private_:bool -> t -> request list -> Tezt_wrapper.JSON.t list Lwt.t

batch_evm_rpc ?private_ evm_node requests sends multiple JSON-RPC requests to the evm_node, for the given requests. If private_ is true, the requests are sent to the private RPC server.

val open_websocket : ?private_:bool -> t -> Tezt_tezos.Websocket.t Lwt.t

Open a websocket connection with the EVM node. If private_ is true, a connection is created with the private websocket endpoint of the node.

val call_evm_websocket : Tezt_tezos.Websocket.t -> request -> Tezt_wrapper.JSON.t Lwt.t

call_evm_websocket ws request sends a JSON-RPC request on the websocket connection ws and waits for the response.

val batch_evm_websocket : Tezt_tezos.Websocket.t -> request list -> Tezt_wrapper.JSON.t list Lwt.t

batch_evm_websocket ws requests sends multiple JSON-RPC requests on the websocket connection ws without waiting for the responses, then receive all responses at the end.

val jsonrpc : ?websocket:Tezt_tezos.Websocket.t -> ?private_:bool -> t -> request -> Tezt_wrapper.JSON.t Lwt.t

jsonrpc uses the websocket to make a JSON-RPC call if provided or falls back to using HTTP RPC request on the EVM node otherwise.

val batch_jsonrpc : ?websocket:Tezt_tezos.Websocket.t -> ?private_:bool -> t -> request list -> Tezt_wrapper.JSON.t list Lwt.t

batch_jsonrpc uses the websocket to make a JSON-RPC calls if provided or falls back to using an HTTP RPC batch request on the EVM node otherwise.

val extract_result : Tezt_wrapper.JSON.t -> Tezt_wrapper.JSON.t

extract_result json expects a JSON-RPC `result` and returns the value.

val extract_error_message : Tezt_wrapper.JSON.t -> Tezt_wrapper.JSON.t

extract_error_message json expects a JSON-RPC `error.message` and returns the value.

val fetch_contract_code : t -> string -> string Lwt.t

fetch_contract_code evm_node contract returns the code associated to the given contract in the rollup.

val upgrade_payload : root_hash:string -> activation_timestamp:string -> string Lwt.t

upgrade_payload ~root_hash ~activation_timestamp gives the upgrade payload to put in a upgrade message, it will upgrade to root_hash at the first block after activation_timestamp (in RFC3399 format).

val sequencer_upgrade_payload : ?client:Tezt_tezos.Client.t -> public_key:string -> pool_address:string -> activation_timestamp:string -> unit -> string Lwt.t

sequencer_upgrade_payload ?client ~public_key ~pool_address ~activation_timestamp () gives the sequencer upgrade payload to put in a upgrade message, it will upgrade the sequencer to public_key at the first l1 block after activation_timestamp (in RFC3399 format).

val init_from_rollup_node_data_dir : ?omit_delayed_tx_events:bool -> t -> Tezt_tezos.Sc_rollup_node.t -> unit Lwt.t

init_from_rollup_node_data_dir ?omit_delayed_tx_events evm_node rollup_node initialises the data dir of the evm node by importing the evm state from a rollup node data dir.

val transform_dump : dump_json:string -> dump_rlp:string -> unit Lwt.t

transform_dump ~dump_json ~dump_rlp transforms a JSON list of instructions stored in dump_json to an RLP list, which is stored in dump_rlp.

val reset : t -> l2_level:int -> unit Lwt.t

reset evm_node ~l2_level reset the store of the evm_node to l2_level.

val chunk_data : rollup_address:string -> ?sequencer_key:string -> ?timestamp:string -> ?parent_hash:string -> ?number:int -> ?client:Tezt_tezos.Client.t -> string list -> string list Lwt.t

chunk data ~rollup_address ?sequencer_key ?timestamp ?parent_hash ?number ?client data generates the valid inputs for the rollup at rollup_address from the given data. If sequencer_key is given, the data produced is for the sequencer mode.

val patch_kernel : t -> string -> unit Lwt.t

patch_kernel evm_node path modifies the kernel used by evm_node with the kernel stored in the file path. This will fail if the node is running.

val patch_state : t -> key:string -> value:string -> unit Lwt.t

patch_kernel evm_node ~key ~value modifies the state of the evm_node by writing value at key.

val export_snapshot : ?compress_on_the_fly:bool -> t -> (Tezt_wrapper.Process.t, string) Tezt_tezos.Runnable.Syntax.runnable

export_snapshot evm_node exports a snapshot of the evm node in a temporary directory. It returns the path for the produced snapshot file.

val import_snapshot : ?force:bool -> t -> snapshot_file:string -> (Tezt_wrapper.Process.t, unit) Tezt_tezos.Runnable.Syntax.runnable

import_snapshot ?force evm_node ~snapshot_file imports the snapshot snapshot_file in the evm node.

val wait_termination : t -> unit Lwt.t

val make_kernel_installer_config : ?max_delayed_inbox_blueprint_length:int -> ?mainnet_compat:bool -> ?remove_whitelist:bool -> ?kernel_root_hash:string -> ?chain_id:int -> ?bootstrap_balance:Wei.t -> ?bootstrap_accounts:string list -> ?sequencer:string -> ?delayed_bridge:string -> ?ticketer:string -> ?administrator:string -> ?sequencer_governance:string -> ?kernel_governance:string -> ?kernel_security_governance:string -> ?minimum_base_fee_per_gas:Wei.t -> ?da_fee_per_byte:Wei.t -> ?delayed_inbox_timeout:int -> ?delayed_inbox_min_levels:int -> ?sequencer_pool_address:string -> ?maximum_allowed_ticks:int64 -> ?maximum_gas_per_transaction:int64 -> ?max_blueprint_lookahead_in_seconds:int64 -> ?set_account_code:(string * string) list -> ?enable_fa_bridge:bool -> ?enable_dal:bool -> ?dal_slots:int list -> output:string -> unit -> (Tezt_wrapper.Process.t, unit) Tezt_tezos.Runnable.t

make_kernel_installer_config ~output () create the config needed for the evm kernel used by the installer

val debug_print_store_schemas : ?path:string -> ?hooks:Tezt_wrapper.Process_hooks.t -> unit -> unit Lwt.t

val man : ?path:string -> ?hooks:Tezt_wrapper.Process_hooks.t -> unit -> unit Lwt.t

val describe_config : ?path:string -> ?hooks:Tezt_wrapper.Process_hooks.t -> unit -> unit Lwt.t

val mode_with_new_private_rpc : mode -> mode

Returns the mode with a fresh private RPC port if one was present.