Module Evm_node.Config_file

Node configuration files.

val read : t -> Tezt_wrapper.JSON.t Lwt.t

Read the configuration file (config.json) of a node.

val write : t -> Tezt_wrapper.JSON.t -> unit Lwt.t

Write the configuration file of a node, replacing the existing one.

val update : t -> (Tezt_wrapper.JSON.t -> Tezt_wrapper.JSON.t) -> unit Lwt.t

Update the configuration file of a node. If the node is already running, it needs to be restarted manually.

Example: Evm_node.Config_file.update evm_node (JSON.put ("experimental_features", experimental_feature))