Module Tezt_etherlink.Rpc

type error = {

code : int;

message : string;

data : string option;

}

type block_param =

| Earliest

| Latest

| Pending

| Number of int
(*
Put directly the value.
*)

| Block_number of {

number : int;

require_canonical : bool;

}
(*
Instead of Number, it uses a JSON object.
*)

| Block_hash of {

hash : string;

require_canonical : bool;

}

module Request : sig ... end

RPC calls wrappers

Calls below are made using an HTTP request to the EVM node by default unless a websocket is provided, in which case the communication happens on it.

val net_version : ?websocket:Tezt_tezos.Websocket.t -> Evm_node.t -> (string, error) Stdlib.result Lwt.t

net_version evm_node calls net_version.

val get_chain_id : ?websocket:Tezt_tezos.Websocket.t -> Evm_node.t -> (int, error) Stdlib.result Lwt.t

get_chain_id evm_node calls eth_getChainId.

val get_transaction_by_hash : ?websocket:Tezt_tezos.Websocket.t -> transaction_hash:string -> Evm_node.t -> (Transaction.transaction_object option, error) Stdlib.result Lwt.t

get_transaction_by_hash ~transaction_hash evm_node calls eth_getTransactionByHash.

val get_code : ?websocket:Tezt_tezos.Websocket.t -> address:string -> Evm_node.t -> (string, error) Stdlib.result Lwt.t

get_code ~address evm_node calls eth_getCode.

val block_number : ?websocket:Tezt_tezos.Websocket.t -> Evm_node.t -> (int32, error) Stdlib.result Lwt.t

block_number evm_node calls eth_blockNumber.

val block_number_opt : ?websocket:Tezt_tezos.Websocket.t -> Evm_node.t -> (int32 option, error) Stdlib.result Lwt.t

block_number_opt evm_node calls eth_blockNumber. allows None when no block have been produced yet.

val get_block_by_number : ?websocket:Tezt_tezos.Websocket.t -> ?full_tx_objects:bool -> block:string -> Evm_node.t -> (Block.t, error) Stdlib.result Lwt.t

get_block_by_number ?full_tx_objets ~block evm_node calls eth_getBlockByNumber. full_tx_objects is false by default, so the block contains the transaction hashes. block can be "latest" or its number.

val get_block_by_hash : ?websocket:Tezt_tezos.Websocket.t -> ?full_tx_objects:bool -> block:string -> Evm_node.t -> (Block.t, error) Stdlib.result Lwt.t

Same as get_block_by_number but uses the hash instead of the number.

val get_gas_price : ?websocket:Tezt_tezos.Websocket.t -> Evm_node.t -> Stdlib.Int32.t Lwt.t

val subscribe : ?websocket:Tezt_tezos.Websocket.t -> kind:Request.subscription_kind -> Evm_node.t -> string Lwt.t

val unsubscribe : ?websocket:Tezt_tezos.Websocket.t -> id:string -> Evm_node.t -> bool Lwt.t

module Syntax : sig ... end

val produce_block : ?websocket:Tezt_tezos.Websocket.t -> ?with_delayed_transactions:bool -> ?timestamp:string -> Evm_node.t -> (int, error) Stdlib.result Lwt.t

produce_block ?with_delayed_transactions ?timestamp evm_node calls the private RPC produceBlock. If provided the block will have timestamp timestamp (in RFC3339) format.

val produce_proposal : ?websocket:Tezt_tezos.Websocket.t -> ?timestamp:string -> Evm_node.t -> (unit, error) Stdlib.result Lwt.t

produce_proposal ?timestamp evm_node calls the private RPC produceProposal. If provided the block will have timestamp timestamp (in RFC3339) format.

val state_value : ?websocket:Tezt_tezos.Websocket.t -> Evm_node.t -> ?block:string -> string -> (string option, error) Stdlib.result Lwt.t

state_value ?websocket evm_node ?block path calls the private RPC stateValue.

val state_subkeys : ?websocket:Tezt_tezos.Websocket.t -> Evm_node.t -> ?block:string -> string -> (string list option, error) Stdlib.result Lwt.t

state_subkeys ?websocket evm_node ?block path calls the private RPC stateSubkeys.

val send_raw_transaction : ?websocket:Tezt_tezos.Websocket.t -> raw_tx:string -> Evm_node.t -> (string, error) Stdlib.result Lwt.t

send_raw_transaction ~raw_tx evm_node calls eth_sendRawTransaction with raw_tx as argument.

val get_transaction_receipt : ?websocket:Tezt_tezos.Websocket.t -> tx_hash:string -> Evm_node.t -> (Transaction.transaction_receipt option, error) Stdlib.result Lwt.t

get_transaction_receipt ~tx_hash evm_node calls eth_getTransactionReceipt with tx_hash as argument.

val estimate_gas : ?websocket:Tezt_tezos.Websocket.t -> (string * Ezjsonm.value) list -> Evm_node.t -> (int64, error) Stdlib.result Lwt.t

estimate_gas eth_call evm_node calls eth_estimateGas with eth_call as payload.

val get_transaction_count : ?websocket:Tezt_tezos.Websocket.t -> ?block:string -> address:string -> Evm_node.t -> (int64, error) Stdlib.result Lwt.t

get_transaction_count ?block ~address evm_node calls eth_getTransactionCount with address as argument (on block, default to "latest" if omitted).

val tez_kernelVersion : ?websocket:Tezt_tezos.Websocket.t -> Evm_node.t -> (string, error) Stdlib.result Lwt.t

tez_kernelVersion evm_node calls tez_kernelVersion. Returns the kernel commit hash.

val tez_kernelRootHash : ?websocket:Tezt_tezos.Websocket.t -> Evm_node.t -> (string option, error) Stdlib.result Lwt.t

tez_kernelRootHash evm_node calls tez_kernelRootHash. Returns the kernel root hash.

val call : ?websocket:Tezt_tezos.Websocket.t -> to_:string -> data:string -> ?block:block_param -> Evm_node.t -> (string, error) Stdlib.result Lwt.t

call ~to_ ~data ?block evm_node call eth_call with to and data as argument. block defaults to Latest.

val get_balance : ?websocket:Tezt_tezos.Websocket.t -> address:string -> ?block:block_param -> Evm_node.t -> (Wei.t, error) Stdlib.result Lwt.t

get_balance ~address ?block evm_node calls eth_getBalance. block defaults to Latest.

val get_storage_at : ?websocket:Tezt_tezos.Websocket.t -> address:string -> ?block:block_param -> pos:string -> Evm_node.t -> (string, error) Stdlib.result Lwt.t

get_balance ~address ?block ~pos evm_node calls eth_getStorageAt. block defaults to Latest.

val get_max_priority_fee_per_gas : ?websocket:Tezt_tezos.Websocket.t -> Evm_node.t -> Stdlib.Int32.t Lwt.t

val replay_block : ?websocket:Tezt_tezos.Websocket.t -> int -> Evm_node.t -> (Block.t, error) Stdlib.result Lwt.t

replay_block number evm_node replays the block number and returns its representation.

type txpool_slot = {

address : string;

transactions : (int64 * Tezt_wrapper.JSON.t) list;

}

A slot in the transaction pool associates an address to a mapping of nonces to transactions.

val txpool_content : ?websocket:Tezt_tezos.Websocket.t -> Evm_node.t -> (txpool_slot list * txpool_slot list, error) Stdlib.result Lwt.t

txpool_content evm_node returns the transaction hash and nonce contained in the `pending` and `queued` pools.

val trace_transaction : ?websocket:Tezt_tezos.Websocket.t -> transaction_hash:string -> ?tracer:string -> ?tracer_config:(string * Tezt_wrapper.JSON.u) list -> Evm_node.t -> (Tezt_wrapper.JSON.t, error) Stdlib.result Lwt.t

trace_transaction ~transaction_hash evm_node replays the given transaction in the same execution context. Doesn't return the trace for now.

val trace_call : ?websocket:Tezt_tezos.Websocket.t -> block:block_param -> to_:string -> data:string -> ?tracer:string -> ?tracer_config:(string * Tezt_wrapper.JSON.u) list -> Evm_node.t -> (Tezt_wrapper.JSON.t, error) Stdlib.result Lwt.t

type fee_history = {

oldest_block : int64;

base_fee_per_gas : int64 list;

gas_used_ratio : float list;

}

val fee_history : ?websocket:Tezt_tezos.Websocket.t -> string -> string -> Evm_node.t -> (fee_history, error) Stdlib.result Lwt.t

fee_history block_count newest_block evm_node calls eth_feeHistory.

val coinbase : ?websocket:Tezt_tezos.Websocket.t -> Evm_node.t -> (string, error) Stdlib.result Lwt.t

coinbase calls eth_coinbase.

val configuration : Evm_node.t -> Tezt_wrapper.JSON.t Lwt.t

Returns the EVM node configuration.