Module Tezt_etherlink.Setup

type l1_contracts = {

delayed_transaction_bridge : string;

exchanger : string;

bridge : string;

admin : string;

sequencer_governance : string;

ticket_router_tester : string;

}

type sequencer_setup = {

node : Tezt_tezos.Node.t;

client : Tezt_tezos.Client.t;

sc_rollup_address : string;

sc_rollup_node : Tezt_tezos.Sc_rollup_node.t;

observer : Evm_node.t;

sequencer : Evm_node.t;

proxy : Evm_node.t;

l1_contracts : l1_contracts;

boot_sector : string;

kernel : Tezt_wrapper.Uses.t;

enable_dal : bool;

}

val uses : Tezt_tezos.Protocol.t -> Tezt_wrapper.Uses.t list

uses protocol returns the list of dependencies for the tests.

val setup_l1_contracts : ?dictator:Tezt_tezos.Account.key -> kernel:Kernel.t -> Tezt_tezos.Client.t -> l1_contracts Lwt.t

setup_l1_contracts ~dictator ~kernel client setups the necessary contracts for the rollup.

val run_new_rpc_endpoint : Evm_node.t -> Evm_node.t Lwt.t

run_new_rpc_endpoint node starts a new rpc node following the setup.

val run_new_observer_node : ?finalized_view:bool -> ?patch_config:(Tezt_wrapper.JSON.t -> Tezt_wrapper.JSON.t) -> sc_rollup_node:Tezt_tezos.Sc_rollup_node.t -> ?rpc_server:Evm_node.rpc_server -> ?websockets:bool -> Evm_node.t -> Evm_node.t Lwt.t

run_new_rpc_endpoint ~sc_rollup_node node starts a new observer following the setup.

val register_test : __FILE__:string -> ?max_delayed_inbox_blueprint_length:int -> ?block_storage_sqlite3:bool -> ?sequencer_rpc_port:int -> ?sequencer_private_rpc_port:int -> ?genesis_timestamp:Tezt_tezos.Client.timestamp -> ?time_between_blocks:Evm_node.time_between_blocks -> ?max_blueprints_lag:int -> ?max_blueprints_ahead:int -> ?max_blueprints_catchup:int -> ?catchup_cooldown:int -> ?delayed_inbox_timeout:int -> ?delayed_inbox_min_levels:int -> ?max_number_of_chunks:int -> ?bootstrap_accounts:string list -> ?sequencer:Tezt_tezos.Account.key -> ?sequencer_pool_address:string -> kernel:Kernel.t -> ?da_fee:Wei.t -> ?minimum_base_fee_per_gas:Wei.t -> ?preimages_dir:string -> ?maximum_allowed_ticks:int64 -> ?maximum_gas_per_transaction:int64 -> ?max_blueprint_lookahead_in_seconds:int64 -> ?enable_fa_bridge:bool -> ?commitment_period:int -> ?challenge_window:int -> ?threshold_encryption:bool -> ?uses:(Tezt_tezos.Protocol.t -> Tezt_wrapper.Uses.t list) -> ?additional_uses:Tezt_wrapper.Uses.t list -> ?history_mode:Tezt_tezos.Sc_rollup_node.history_mode -> enable_dal:bool -> ?dal_slots:int list option -> ?rpc_server:Evm_node.rpc_server -> ?websockets:bool -> (sequencer_setup -> Tezt_tezos.Protocol.t -> unit Lwt.t) -> title:string -> tags:string list -> Tezt_tezos.Protocol.t list -> unit

register_test setups the full sequencer environment for the tests and starts the test. For each of the parameters, please refer to Evm_node.

val register_test_for_kernels : __FILE__:string -> ?max_delayed_inbox_blueprint_length:int -> ?block_storage_sqlite3:bool -> ?sequencer_rpc_port:int -> ?sequencer_private_rpc_port:int -> ?genesis_timestamp:Tezt_tezos.Client.timestamp -> ?time_between_blocks:Evm_node.time_between_blocks -> ?max_blueprints_lag:int -> ?max_blueprints_ahead:int -> ?max_blueprints_catchup:int -> ?catchup_cooldown:int -> ?delayed_inbox_timeout:int -> ?delayed_inbox_min_levels:int -> ?max_number_of_chunks:int -> ?bootstrap_accounts:string list -> ?sequencer:Tezt_tezos.Account.key -> ?sequencer_pool_address:string -> ?kernels:Kernel.t list -> ?da_fee:Wei.t -> ?minimum_base_fee_per_gas:Wei.t -> ?preimages_dir:string -> ?maximum_allowed_ticks:int64 -> ?maximum_gas_per_transaction:int64 -> ?max_blueprint_lookahead_in_seconds:int64 -> ?enable_fa_bridge:bool -> ?history_mode:Tezt_tezos.Sc_rollup_node.history_mode -> ?commitment_period:int -> ?challenge_window:int -> ?additional_uses:Tezt_wrapper.Uses.t list -> threshold_encryption:bool -> enable_dal:bool -> ?dal_slots:int list option -> ?rpc_server:Evm_node.rpc_server -> ?websockets:bool -> title:string -> tags:string list -> (sequencer_setup -> Tezt_tezos.Protocol.t -> unit Lwt.t) -> Tezt_tezos.Protocol.t list -> unit

register_test_for_kernels setups the full sequencer environment for the tests and starts the test. For each of the parameters, please refer to Evm_node. The test is registered for each kernel.

val setup_sequencer : ?max_delayed_inbox_blueprint_length:int -> ?next_wasm_runtime:bool -> ?block_storage_sqlite3:bool -> ?sequencer_rpc_port:int -> ?sequencer_private_rpc_port:int -> mainnet_compat:bool -> ?genesis_timestamp:Tezt_tezos.Client.timestamp -> ?time_between_blocks:Evm_node.time_between_blocks -> ?max_blueprints_lag:int -> ?max_blueprints_ahead:int -> ?max_blueprints_catchup:int -> ?catchup_cooldown:int -> ?delayed_inbox_timeout:int -> ?delayed_inbox_min_levels:int -> ?max_number_of_chunks:int -> ?commitment_period:int -> ?challenge_window:int -> ?bootstrap_accounts:string list -> ?sequencer:Tezt_tezos.Account.key -> ?sequencer_pool_address:string -> ?kernel:Tezt_wrapper.Uses.t -> ?da_fee:Wei.t -> ?minimum_base_fee_per_gas:Wei.t -> ?preimages_dir:string -> ?maximum_allowed_ticks:int64 -> ?maximum_gas_per_transaction:int64 -> ?max_blueprint_lookahead_in_seconds:int64 -> ?enable_fa_bridge:bool -> ?threshold_encryption:bool -> ?drop_duplicate_when_injection:bool -> ?blueprints_publisher_order_enabled:bool -> ?history_mode:Tezt_tezos.Sc_rollup_node.history_mode -> enable_dal:bool -> ?dal_slots:int list -> ?rpc_server:Evm_node.rpc_server -> ?websockets:bool -> Tezt_tezos.Protocol.t -> sequencer_setup Lwt.t