Module Tezt_etherlink.Wei

Helpers for dealing with units of wei. `wei` is the smallest denomination of ether. 1 eth = 1_000_000_000_000_000_000 wei (10^18)

Please note that none of the functions here perform any bounds checks.

type t

A unit of wei

val zero : t

0 wei

val one : t

1 wei

val one_eth : t

1 eth, or 10^18 wei

val to_string : t -> string

Convert t to a string.

val of_string : string -> t

Convert a string to a t.

val to_wei_z : Z.t -> t

Converts a Z.t to a wei value.

val of_wei_z : t -> Z.t

Converts a wei value to a Z.t.

val of_tez : Tezt_tezos.Tez.t -> t

Converts a tez value to the equivalent wei. *

val of_eth_int : int -> t

Convert an int amount of eth to a wei one. This doesn't perform any bounds checks.

val of_eth_string : string -> t

Convert a string amount of eth to a wei one. This doesn't perform any bounds checks. It assumes the string represents either an integer or a float - both denoting an amount of eth.

val of_gwei_string : string -> t

Convert a string amount of gwei to a wei one. This doesn't perform any bounds checks. It assumes the string represents either an integer or a float - both denoting an amount of gwei.

val (+) : t -> t -> t

Addition. This doesn't perform any bounds checks.

val (-) : t -> t -> t

Subtraction. This doesn't perform any bound checks.

val (*) : t -> Z.t -> t

Multiplication. This doesn't perform any bound checks.

val (/) : t -> Z.t -> t

Division. This doesn't perform any bound checks.

val cdiv : t -> Z.t -> t

Division, rounding towards +oo.

val typ : t Tezt_wrapper.Check.typ

The wei Check.typ.

val to_le_bytes : t -> Stdlib.Bytes.t

Convert t to U256 little endian bytes.

val truncate_to_mutez : t -> int

Truncate the Wei (10^18) to Mutez (10^6).